requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
asgiref>=3.8.1
sqlparse>=0.5.3
typing-extensions>=4.13.2
tzdata>=2025.2 